tests: Move TestCrossRealmPMs to test_message_send.py.
